requirements.txt
fastapi
uvicorn
livekit-agents
livekit-plugins-google
livekit-api
google-generativeai
python-dotenv